requirements.txt
diffusers>=0.21.0
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
pyzbar>=0.1.9
opencv-python>=4.8.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

